Timer Implemented
Timer displays on Game JSP and decrements every second. Uses a script based on one found on W3Schools.
Able to pass timer value with user input, allowing time to be passed to the Game class. However, this leads to the doGet method resetting the timer to the last time set by the doPost method.
Timer is not yet triggering any game events.
Displays time in Hours / Minutes / Seconds