Implemented Login, Added fields and constructor to FakeDB
Server redirects user if they are not logged in. Two logins are currently hard-coded into the TitlePageServlet.
Test file created for FakeDB, along with constructor and fields for the FakeDB so that we can begin working on queries.
Made it so the TitlePage loads Jordan's clock gif.
User can log in from the title page via an overlay similar to the character creation overlay.