Fixed issue with Player and Game not saving
Resolved issue with Game and Player not saving properly. Needed to write a query in updateGameByGameId that updated the Game fields.
Player lost its ID due to the former process of character creation creating a new Player object instead of using setters on the existing Player. Hence, it couldn't save Player data without the correct ID.