LevelUp System part 1
I have the overlay made for the levelup prompt when the player eventually gets to the threshold. I also have a simple bar representing the xp progress. I hope to have it move when xp is gained. Next, I have the command to give the player xp to test these functions.