Level up System part 2
I have developed the original features from before. For the overlay I made sure that is reliably closes and returns when the threshold is met. In addition, the overlay now updates the players stats where the player allocates their points. As a result, I have added some features from the Character creation like the error system so points are allocated correctly. Lastly, the xp gain is still very simple and hasn't been developed enough where it relies on the DB. However, it is now rewarding the player from doing various actions.